// Tarea1/src/app/models/perito.ts
export class Perito {
    dniPerito: number = 0;
    nombrePerito: string = '';
    apellidoPerito1: string = '';
    apellidoPerito2: string = '';
    telefonoContacto: string = '';
    telefonoOficina: string = '';
    claseVia: string = '';
    nombreVia: string = '';
    numeroVia: string = '';
    codPostal: string = '';
    ciudad: string = '';

    constructor() {
        this.dniPerito = 0;
        this.nombrePerito = '';
        this.apellidoPerito1 = '';
        this.apellidoPerito2 = '';
        this.telefonoContacto = '';
        this.telefonoOficina = '';
        this.claseVia = '';
        this.nombreVia = '';
        this.numeroVia = '';
        this.codPostal = '';
        this.ciudad = '';
    }
}